// src/api/model/guestBookResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Forgather API
 * 당신을 위한 순간, 흩어지지 않게. Forgather
 * OpenAPI spec version: 2.0.0
 */
import type { GuestBookCardSimpleResponse } from './guestBookCardSimpleResponse';

export interface GuestBookResponse {
  /** 읽지 않은 방명록 카드 개수 */
  unreadCount?: number;
  /** 방명록 카드 목록 */
  guestBookCards?: GuestBookCardSimpleResponse[];
  /** 현재 페이지 번호 */
  currentPage?: number;
  /** 페이지 당 방명록 카드 개수 */
  pageSize?: number;
  /** 총 방명록 카드 개수 */
  totalCount?: number;
  /** 총 페이지 수 */
  totalPages?: number;
}